requirements.txt
python-telegram-bot
pysoundfile
